Fix a bug on the way canvas & matrix are working.
The printMatrix function was displaying the candidate rotated because it was printing
the m[i] then m[j] while it should be reversed.

In a similar way the symetry axis was on axis 'Y' while in the sofa simulation it
was on the 'X' axis.

Now everything should be consistant.